crimtech_final_project/requirements.txt
BeautifulSoup<4.0
boto==2.46.1
boto3==1.5.13
celery==3.1.18
Django==1.8.18
django-bower==5.0.4
django-celery==3.1.17
django-compressor==2.2
django-debug-toolbar==1.4
django-downloadview==1.8
django-extensions==1.5.9
django-libsass==0.6
django-localflavor-us==1.1
django-model-utils==2.3.1
django-select2==4.3.1
django-solo==1.1.0
django-sortedm2m==1.0.2
django-storages==1.1.8
django-wysiwyg-redactor==0.4.9
easy-thumbnails==2.2.1
Fabric==1.10.2
gdata==2.0.18
google-api-python-client==1.4.2
httplib2==0.9.2
kombu==3.0.30
mailchimp==2.0.9
mysqlclient==1.3.7
newrelic==2.58.1.44
paramiko==1.10.0
Pillow==3.0.0
premailer==2.9.6
pycrypto==2.6.1
python-dateutil==2.6.1
python-memcached==1.57
python-social-auth==0.2.3
pytz==2016.7
raven==5.8.1
requests==2.18.4
six==1.11.0
sphinx==1.3.1
supervisor==3.1.3
wsgiref==0.1.2
